Refactor param types
