+++ lib/blade/button/plugin.ts
@@ -1,6 +1,6 @@
 import {ValueMap} from '../../common/model/value-map';
+import {BladeParams} from '../../common/params';
 import {ParamsParsers, parseParams} from '../../common/params-parsers';
-import {BladeParams} from '../common/api/params';
 import {LabelController} from '../label/controller/label';
 import {BladePlugin} from '../plugin';
 import {ButtonApi} from './api/button';
