+++ lib/blade/common/api/blade-rack.ts
@@ -1,3 +1,4 @@
+import {BladeParams} from '../../../common/params';
 import {View} from '../../../common/view/view';
 import {ButtonApi} from '../../button/api/button';
 import {FolderApi} from '../../folder/api/folder';
@@ -8,7 +9,6 @@ import {TabApi} from '../../tab/api/tab';
 import {BladeController} from '../controller/blade';
 import {BladeApi} from './blade';
 import {
-	BladeParams,
 	ButtonParams,
 	FolderParams,
 	InputParams,
