+++ lib/blade/common/api/plugins.ts
@@ -1,4 +1,5 @@
 import {BindingTarget} from '../../../common/binding/target';
+import {BaseInputParams, BaseMonitorParams} from '../../../common/params';
 import {TpError} from '../../../common/tp-error';
 import {View} from '../../../common/view/view';
 import {
@@ -20,12 +21,7 @@ import {RackApi} from '../../rack/api/rack';
 import {RackController} from '../../rack/controller/rack';
 import {BladeController} from '../controller/blade';
 import {BladeApi} from './blade';
-import {
-	BaseInputParams,
-	BaseMonitorParams,
-	InputParams,
-	MonitorParams,
-} from './params';
+import {InputParams, MonitorParams} from './params';
 
 export const Plugins: {
 	blades: BladePlugin<any>[];
