+++ lib/blade/list/plugin.ts
@@ -1,13 +1,14 @@
 import {ListController} from '../../common/controller/list';
 import {ValueMap} from '../../common/model/value-map';
 import {createValue} from '../../common/model/values';
+import {ListParamsOptions} from '../../common/params';
+import {BladeParams} from '../../common/params';
 import {
 	ParamsParser,
 	ParamsParsers,
 	parseParams,
 } from '../../common/params-parsers';
 import {normalizeListOptions, parseListOptions} from '../../common/util';
-import {BladeParams, ListParamsOptions} from '../common/api/params';
 import {LabeledValueController} from '../label/controller/value-label';
 import {BladePlugin} from '../plugin';
 import {ListApi} from './api/list';
