+++ lib/input-binding/point-2d/view/point-2d-picker.ts
@@ -1,8 +1,8 @@
-import {PickerLayout} from '../../../blade/common/api/params';
 import {SVG_NS} from '../../../common/dom-util';
 import {Value} from '../../../common/model/value';
 import {ViewProps} from '../../../common/model/view-props';
 import {mapRange} from '../../../common/number-util';
+import {PickerLayout} from '../../../common/params';
 import {ClassName} from '../../../common/view/class-name';
 import {View} from '../../../common/view/view';
 import {Point2d} from '../model/point-2d';
