+++ lib/input-binding/point-2d/view/point-2d.ts
@@ -1,8 +1,8 @@
-import {PickerLayout} from '../../../blade/common/api/params';
 import {createSvgIconElement} from '../../../common/dom-util';
 import {bindValue} from '../../../common/model/reactive';
 import {Value} from '../../../common/model/value';
 import {ViewProps} from '../../../common/model/view-props';
+import {PickerLayout} from '../../../common/params';
 import {ClassName} from '../../../common/view/class-name';
 import {valueToClassName} from '../../../common/view/reactive';
 import {View} from '../../../common/view/view';
