+++ lib/monitor-binding/string/plugin.ts
@@ -1,12 +1,12 @@
-import {BaseMonitorParams} from '../../blade/common/api/params';
 import {formatString, stringFromUnknown} from '../../common/converter/string';
+import {BaseMonitorParams} from '../../common/params';
 import {ParamsParsers, parseParams} from '../../common/params-parsers';
 import {Constants} from '../../misc/constants';
 import {MultiLogController} from '../common/controller/multi-log';
 import {SingleLogMonitorController} from '../common/controller/single-log';
 import {MonitorBindingPlugin} from '../plugin';
 
-interface StringMonitorParams extends BaseMonitorParams {
+export interface StringMonitorParams extends BaseMonitorParams {
 	lineCount?: number;
 	multiline?: boolean;
 }
