+++ test-module/plugin/src/index.ts
@@ -1,10 +1,10 @@
 import {Tweakpane} from 'tweakpane';
-import {BaseInputParams} from 'tweakpane/lib/blade/common/api/params';
 import {Controller} from 'tweakpane/lib/common/controller/controller';
 // Import individual classes
 import {stringFromUnknown} from 'tweakpane/lib/common/converter/string';
 import {Value} from 'tweakpane/lib/common/model/value';
 import {ViewProps} from 'tweakpane/lib/common/model/view-props';
+import {BaseInputParams} from 'tweakpane/lib/common/params';
 import {writePrimitive} from 'tweakpane/lib/common/primitive';
 import {ClassName} from 'tweakpane/lib/common/view/class-name';
 import {View} from 'tweakpane/lib/common/view/view';
